+++ src/Text/XML/DOM/Parser.hs
@@ -1,9 +1,13 @@
 module Text.XML.DOM.Parser
-  ( module Text.XML.DOM.Parser.Class
-  , module Text.XML.DOM.Parser.Combinators
+  ( module Text.XML.DOM.Parser.Combinators
+  , module Text.XML.DOM.Parser.Common
+  , module Text.XML.DOM.Parser.FromAttribute
+  , module Text.XML.DOM.Parser.FromDom
   , module Text.XML.DOM.Parser.Types
   ) where
 
-import Text.XML.DOM.Parser.Class
 import Text.XML.DOM.Parser.Combinators
+import Text.XML.DOM.Parser.Common
+import Text.XML.DOM.Parser.FromAttribute
+import Text.XML.DOM.Parser.FromDom
 import Text.XML.DOM.Parser.Types
